requirements.txt
mysql-connector-python
streamlit
pandas
matplotlib
scikit-learn
bcrypt
python-dotenv
